pyproject.toml
[project]
name = "vocode-main"
version = "0.1.0"
description = ""
authors = [
    {name = "Mellettan",email = "roma23470@gmail.com"}
]

requires-python = ">=3.11 <4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "langchain-gigachat (>=0.3.4,<0.4.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "grpcio-tools (>=1.70.0,<2.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "sounddevice (>=0.5.1,<0.6.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "websockets (>=15.0,<16.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "redis (>=5.2.1,<6.0.0)",
    "black (>=25.1.0,<26.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
